social_media_api/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
Pillow==11.0.0
requests==2.32.3
